Add code for early initialization.
Not all devices can be initialize early. RAM-based black boxes should
generally be supportable, but may need special attention to work with
ECC. So, instead of just zeroing all of memory in order to set the ECC
bits properly, it will be necessary to read and then write the data.
Even this may not work with some devices. Other devices, such as
NAND-based devices, will probably require more effort, but registering
such devices with the driver so that the black box code is called
immediately after the underlying device is initialized should get them
working as early as possible.